+++ package.json
@@ -4,7 +4,9 @@
   "description": "TicTacToe",
   "main": "tictac.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest",
+    "start:dev": "webpack-dev-server --mode development --open",
+    "start": "webpack"
   },
   "repository": {
     "type": "git",
@@ -16,5 +18,18 @@
   "bugs": {
     "url": "https://github.com/lucasmazo32/jstic-tac-toe/issues"
   },
-  "homepage": "https://github.com/lucasmazo32/jstic-tac-toe#readme"
+  "homepage": "https://github.com/lucasmazo32/jstic-tac-toe#readme",
+  "devDependencies": {
+    "@babel/preset-env": "^7.9.5",
+    "clean-webpack-plugin": "^3.0.0",
+    "css-loader": "^3.5.1",
+    "html-webpack-plugin": "^4.2.0",
+    "jest": "^25.3.0",
+    "mini-css-extract-plugin": "^0.9.0",
+    "node-sass": "^4.13.1",
+    "sass-loader": "^8.0.2",
+    "webpack": "^4.42.1",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3"
+  }
 }
