+++ src/domChange.js
@@ -1,28 +1,4 @@
-const Player = (name, icon) => ({ name, icon });
-
-const Board = () => {
-  const board = ['', '', '', '', '', '', '', '', ''];
-  const move = (pos, icon) => {
-    board[pos] = icon;
-  };
-  const empty = () => {
-    for (let index = 0; index < board.length; index += 1) {
-      board[index] = '';
-    }
-  };
-  return { board, move, empty };
-};
-
-const GameFlow = (game) => {
-  const move = (pos, icon) => {
-    if (game.board[pos] === '') {
-      game.move(pos, icon);
-      return true;
-    }
-    return false;
-  };
-  return { move };
-};
+import { Player, Board, GameFlow } from './ticSet';
 
 // Animations
 
@@ -38,80 +14,82 @@ const inputContainer = document.querySelector('.input-container');
 const winnerField = document.querySelector('.winner-field');
 const winnerDiv = document.querySelector('.winner');
 const playAgain = document.querySelector('#play-again');
-
-const InputCheck = (input1, input2) => {
-  const boxCheck = event => {
-    event.addEventListener('input', () => {
-      if (event.value.length !== 0 && event.value.length < 3) {
-        event.style.boxShadow = '0px 0px 5px 2px red';
-      } else {
-        event.style.boxShadow = '';
-      }
-    });
-  };
-  boxCheck(input1);
-  boxCheck(input2);
-  const readyTG = (event1, event2) => {
-    event1.oninput = () => {
-      if (event1.value.length >= 3 && event2.value.length >= 3) {
-        playBtn.style.opacity = '1';
-        playBtn.style.pointerEvents = 'inherit';
-      } else {
-        playBtn.style.opacity = '';
-        playBtn.style.pointerEvents = '';
-      }
-    };
-    event2.oninput = () => {
-      if (event1.value.length >= 3 && event2.value.length >= 3) {
-        playBtn.style.opacity = '1';
-        playBtn.style.pointerEvents = 'inherit';
-      } else {
-        playBtn.style.opacity = '';
-        playBtn.style.pointerEvents = '';
-      }
-    };
-  };
-  readyTG(input1, input2);
-};
-
-InputCheck(player1Input, player2Input);
-
-playBtn.onmouseenter = () => {
-  gameContainer.style.opacity = '0.3';
-};
-
 let nClick = 0;
 
-playBtn.onclick = () => {
-  gameContainer.style.opacity = '1';
-  gameContainer.style.pointerEvents = 'inherit';
-  gameContainer.style.marginTop = '0';
+const animations = () => {
+  const InputCheck = (input1, input2) => {
+    const boxCheck = event => {
+      event.addEventListener('input', () => {
+        if (event.value.length !== 0 && event.value.length < 3) {
+          event.style.boxShadow = '0px 0px 5px 2px red';
+        } else {
+          event.style.boxShadow = '';
+        }
+      });
+    };
+    boxCheck(input1);
+    boxCheck(input2);
+    const readyTG = (event1, event2) => {
+      event1.oninput = () => {
+        if (event1.value.length >= 3 && event2.value.length >= 3) {
+          playBtn.style.opacity = '1';
+          playBtn.style.pointerEvents = 'inherit';
+        } else {
+          playBtn.style.opacity = '';
+          playBtn.style.pointerEvents = '';
+        }
+      };
+      event2.oninput = () => {
+        if (event1.value.length >= 3 && event2.value.length >= 3) {
+          playBtn.style.opacity = '1';
+          playBtn.style.pointerEvents = 'inherit';
+        } else {
+          playBtn.style.opacity = '';
+          playBtn.style.pointerEvents = '';
+        }
+      };
+    };
+    readyTG(input1, input2);
+  };
 
-  playersShow.forEach(player => {
-    player.style.opacity = '1';
-  });
+  InputCheck(player1Input, player2Input);
 
-  inputContainer.style.marginTop = '0';
-  inputContainer.style.opacity = '0';
-  inputContainer.style.pointerEvents = 'none';
-  inputContainer.style.flexBasis = '20%';
-  inputContainer.style.minWidth = '20%';
+  playBtn.onmouseenter = () => {
+    gameContainer.style.opacity = '0.3';
+  };
 
-  squares.forEach(square => {
-    square.style.width = '120px';
-    square.style.height = '120px';
-  });
 
-  window.setTimeout(() => {
-    playBtn.style.visibility = 'hidden';
-    inputContainer.style.visibility = 'hidden';
-  }, 1000);
-};
+  playBtn.onclick = () => {
+    gameContainer.style.opacity = '1';
+    gameContainer.style.pointerEvents = 'inherit';
+    gameContainer.style.marginTop = '0';
 
-playBtn.onmouseleave = () => {
-  if (nClick === 0) {
-    gameContainer.style.opacity = '';
-  }
+    playersShow.forEach(player => {
+      player.style.opacity = '1';
+    });
+
+    inputContainer.style.marginTop = '0';
+    inputContainer.style.opacity = '0';
+    inputContainer.style.pointerEvents = 'none';
+    inputContainer.style.flexBasis = '20%';
+    inputContainer.style.minWidth = '20%';
+
+    squares.forEach(square => {
+      square.style.width = '120px';
+      square.style.height = '120px';
+    });
+
+    window.setTimeout(() => {
+      playBtn.style.visibility = 'hidden';
+      inputContainer.style.visibility = 'hidden';
+    }, 1000);
+  };
+
+  playBtn.onmouseleave = () => {
+    if (nClick === 0) {
+      gameContainer.style.opacity = '';
+    }
+  };
 };
 
 const afterWin = places => {
@@ -220,16 +198,20 @@ function gamePlay(player1, player2) {
   });
 }
 
-playBtn.addEventListener('click', () => {
-  playBtn.style.pointerEvents = 'none';
-  playBtn.style.height = '0';
-  playBtn.style.marginTop = '0';
-  playBtn.style.opacity = '0';
+const playGame = () => {
+  playBtn.addEventListener('click', () => {
+    playBtn.style.pointerEvents = 'none';
+    playBtn.style.height = '0';
+    playBtn.style.marginTop = '0';
+    playBtn.style.opacity = '0';
+    
+    player1Name.innerHTML = player1Input.value;
+    player2Name.innerHTML = player2Input.value;
+    const player1 = Player(player1Input.value, 'X');
+    const player2 = Player(player2Input.value, 'O');
+    gamePlay(player1, player2);
+    nClick += 1;
+  });
+};
 
-  player1Name.innerHTML = player1Input.value;
-  player2Name.innerHTML = player2Input.value;
-  const player1 = Player(player1Input.value, 'X');
-  const player2 = Player(player2Input.value, 'O');
-  gamePlay(player1, player2);
-  nClick += 1;
-});
+export { playGame, animations };
