+++ src/index.js
@@ -1,0 +1,5 @@
+import './style.scss';
+import { playGame, animations } from './domChange';
+
+animations();
+playGame();
