+++ src/template.html
@@ -5,7 +5,7 @@
   <link href="https://fonts.googleapis.com/css?family=Baloo+Chettan+2&display=swap" rel="stylesheet">
   <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css" integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">
   <title>Sandbox</title>
-<link href="style.css" rel="stylesheet"></head>
+</head>
 <body>
   <div class="title">
     <h1>Tic Tac Toe</h1>
@@ -56,5 +56,5 @@
     <span class="winner-field"></span>
     <a class="btn btn-dark" id="play-again">Play again!</a>
   </div>
-<script src="main.js"></script></body>
+</body>
 </html>
