+++ test/boardLogic.test.js
@@ -1,0 +1,6 @@
+import { Board, GameFlow } from '../src/index';
+
+test('Board instance', () => {
+  const board = Board();
+  expect(board.board).toBe(['','','','','','','','','']);
+});
